refactor: split modern.ts into focused sub-modules
- Extract ai, kiku, lapis, proxy, tags, fields, media, behavior, metadata into own files
- Move applyAnkiKikuResolution import to kiku.ts
- Fix nPlusOne migration: later invalid duplicate values now correctly supersede earlier valid ones
